fast_api/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-cloud-storage
requests
pydantic
python-multipart
python-dotenv
numpy
tritonclient[all]
gevent
prometheus-fastapi-instrumentator